Cube3D: rename algs to sets and wrap sets in algs
this commit is a bit hard to explain, but essentially, the new structure Cube3 uses is that sets are a set of moves with the same front_face and orient, and algorithms are a group of sets.

in order to avoid changing large amount of Tutorial code, permute() has been changed to append_buffer() to add a set to a "buffer" algorithm, and flush_buffer() sends this algorithm to both Cube and Cube3D.

this should allow the replay of certain steps, rather than individual sets as RCT previously could do.